requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
requests>=2.25.0